requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-cors-headers==3.0.2
django-debug-toolbar==2.0
django-jet==1.0.8
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jsonfield2==3.0.2
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
nltk==3.4.3
numpy==1.16.4
openapi-codec==1.3.2
pytz==2019.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
